+++ examples/become-ansibleplaybook/become-ansibleplaybook.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	ansibler "github.com/apenella/go-ansible"
+	"github.com/apenella/go-ansible/execute"
 )
 
 func main() {
@@ -26,7 +27,9 @@ func main() {
 		ConnectionOptions:          ansiblePlaybookConnectionOptions,
 		PrivilegeEscalationOptions: ansiblePlaybookPrivilegeEscalationOptions,
 		Options:                    ansiblePlaybookOptions,
-		ExecPrefix:                 "Go-ansible example with become",
+		Exec: execute.NewDefaultExecute(
+			execute.WithPrefix("Go-ansible example with become"),
+		),
 	}
 
 	err := playbook.Run(context.TODO())