+++ examples/json-stdout-ansibleplaybook/json-stdout-ansibleplaybook.go
@@ -7,6 +7,7 @@ import (
 	"io"
 
 	ansibler "github.com/apenella/go-ansible"
+	"github.com/apenella/go-ansible/execute"
 	"github.com/apenella/go-ansible/stdoutcallback/results"
 )
 
@@ -25,13 +26,16 @@ func main() {
 		Inventory: "127.0.0.1,",
 	}
 
+	execute := execute.NewDefaultExecute(
+		execute.WithWrite(io.Writer(buff)),
+	)
+
 	playbook := &ansibler.AnsiblePlaybookCmd{
 		Playbook:          "site.yml",
+		Exec:              execute,
 		ConnectionOptions: ansiblePlaybookConnectionOptions,
 		Options:           ansiblePlaybookOptions,
-		ExecPrefix:        "Go-ansible example",
 		StdoutCallback:    "json",
-		Writer:            io.Writer(buff),
 	}
 
 	err = playbook.Run(context.TODO())