+++ examples/myexecutor-ansibleplaybook/myexecutor-ansibleplaybook.go
@@ -5,12 +5,35 @@ import (
 	"fmt"
 
 	ansibler "github.com/apenella/go-ansible"
+	"github.com/apenella/go-ansible/execute"
+	"github.com/apenella/go-ansible/stdoutcallback"
 )
 
-type MyExecutor struct{}
+type MyExecutor struct {
+	Prefix string
+}
 
-func (e *MyExecutor) Execute(ctx context.Context, command string, args []string, prefix string) error {
-	fmt.Println("I am doing nothing")
+func (e *MyExecutor) Options(options ...execute.ExecuteOptions) {
+	// apply all options to the executor
+	for _, opt := range options {
+		opt(e)
+	}
+}
+
+func WithPrefix(prefix string) execute.ExecuteOptions {
+	return func(e execute.Executor) {
+		e.(*MyExecutor).Prefix = prefix
+	}
+}
+
+func (e *MyExecutor) Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...execute.ExecuteOptions) error {
+
+	// apply all options to the executor
+	for _, opt := range options {
+		opt(e)
+	}
+
+	fmt.Println(fmt.Sprintf("%s %s\n", e.Prefix, "I am MyExecutor and I am doing nothing"))
 
 	return nil
 }
@@ -25,11 +48,16 @@ func main() {
 		Inventory: "127.0.0.1,",
 	}
 
+	e := &MyExecutor{}
+	e.Options(
+		WithPrefix("[Go ansible example]"),
+	)
+
 	playbook := &ansibler.AnsiblePlaybookCmd{
 		Playbook:          "site.yml",
 		ConnectionOptions: ansiblePlaybookConnectionOptions,
 		Options:           ansiblePlaybookOptions,
-		Exec:              &MyExecutor{},
+		Exec:              e,
 	}
 
 	err := playbook.Run(context.TODO())