+++ examples/simple-ansibleplaybook/simple-ansibleplaybook.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	ansibler "github.com/apenella/go-ansible"
+	"github.com/apenella/go-ansible/execute"
 )
 
 func main() {
@@ -21,7 +22,9 @@ func main() {
 		Playbook:          "site.yml",
 		ConnectionOptions: ansiblePlaybookConnectionOptions,
 		Options:           ansiblePlaybookOptions,
-		ExecPrefix:        "Go-ansible example",
+		Exec: execute.NewDefaultExecute(
+			execute.WithPrefix("Go-ansible example"),
+		),
 	}
 
 	err := playbook.Run(context.TODO())