+++ execute/defaultExecute_test.go
@@ -36,29 +36,25 @@ func TestDefaultExecute(t *testing.T) {
 		{
 			desc: "Testing an ansible-playbook with local connection",
 			err:  nil,
-			execute: &DefaultExecute{
-				Write:       io.Writer(&stdout),
-				WriterError: io.Writer(&stderr),
-			},
-			ctx:     context.TODO(),
-			command: []string{binary, "--inventory", "127.0.0.1,", "test/site.yml", "-c", "local"},
-			options: []ExecuteOptions{
+			execute: NewDefaultExecute(
+				WithWrite(io.Writer(&stdout)),
+				WithWriteError(io.Writer(&stderr)),
 				WithPrefix("test"),
-			},
+			),
+			ctx:            context.TODO(),
+			command:        []string{binary, "--inventory", "127.0.0.1,", "test/site.yml", "-c", "local"},
 			expectedStdout: ``,
 		},
 		{
 			desc: "Testing an ansible-playbook forcing an invalid charaters warning message",
 			err:  errors.New("(DefaultExecute::Execute)", "Error during command execution: ansible-playbook error: parser error\n\nCommand executed: "+binary+" --inventory test/all test/site.yml --user apenella\n\nexit status 4"),
-			execute: &DefaultExecute{
-				Write:       io.Writer(&stdout),
-				WriterError: io.Writer(&stderr),
-			},
+			execute: NewDefaultExecute(
+				WithWrite(io.Writer(&stdout)),
+				WithWriteError(io.Writer(&stderr)),
+				WithPrefix("test"),
+			),
 			ctx:     context.TODO(),
 			command: []string{binary, "--inventory", "test/all", "test/site.yml", "--user", "apenella"},
-			options: []ExecuteOptions{
-				WithPrefix("test"),
-			},
 			expectedStderr: `test ── [WARNING]: Invalid characters were found in group names but not replaced, use
 test ── -vvvv to see details
 `,
@@ -71,7 +67,7 @@ test ── -vvvv to see details
 		stdout.Reset()
 		stderr.Reset()
 
-		err := test.execute.Execute(test.ctx, test.command, test.options...)
+		err := test.execute.Execute(test.ctx, test.command, nil, test.options...)
 
 		if err != nil && assert.Error(t, err) {
 			assert.Equal(t, test.err, err)