+++ CHANGELOG.md
@@ -14,8 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BREAKING CHANGE**: `Executor` interface has been moved from `ansibler` package to `execute` package.
 - **BREAKING CHANGE**: `Executor` interface is changed to `Execute(ctx context.Context, command []string, options ...ExecuteOptions) error`.
 - **BREAKING CHANGE**: `DefaultExecute` has been updated to use options pattern design, and includes a set of `WithXXX` methods for each attribute.
+
+## Removed
 - **BREAKING CHANGE**: Remove `ExecPrefix` from `AnsiblePlaybookCmd`
 - **BREAKING CHANGE**: Remove `CmdRunDir` from `AnsiblePlaybookCmd`
+- **BREAKING CHANGE**: Remove `Writer` from `AnsiblePlaybookCmd`
 
 ## [v0.8.0]
 ### Added